refactored code to  improve the readability and maintainability
