+++ src/modules/countComments.test.js
@@ -14,6 +14,10 @@ describe('countComments function', () => {
       document.createElement('li'),
       document.createElement('li'),
     ];
+
+    /* (mockReturnValue) method to specify what
+     the mocked function should return when called.
+    */
     document.querySelectorAll.mockReturnValue(mockCommentElements);
 
     const result = countComments();
