+++ src/modules/displayDetailCard.js
@@ -1,6 +1,8 @@
-const showDetail = document.querySelector('.show-detail');
+import { getItemComments } from './commentsApi.js';
+import countComments from './countComments.js';
 
 const displayDetailCard = (data) => {
+  const showDetail = document.querySelector('.show-detail');
   document.getElementById('shows-comment').style.display = 'flex';
   const htmlDetail = `<div class="image-container">
                           <img class="comment-image" src="${data.image.medium}" />
@@ -24,4 +26,33 @@ const displayDetailCard = (data) => {
   showDetail.innerHTML = htmlDetail;
 };
 
-export default displayDetailCard;
+const displayComments = async (commentId) => {
+  const listCommet = document.querySelector('.listComment');
+  const comments = await getItemComments(commentId);
+  listCommet.innerHTML = '';
+
+  comments.forEach((comment) => {
+    const li = document.createElement('li');
+    li.classList.add('itemComment');
+    li.innerHTML = `Date: ${comment.creation_date} <strong>${comment.username}</strong>  : ${comment.comment} `;
+    listCommet.appendChild(li);
+  });
+};
+
+const displayCounterOfComments = () => {
+  const numberCommet = document.querySelector('.numberCommet');
+  // Wait a bit to allow comments to be displayed
+  setTimeout(() => {
+    const commentCounter = countComments();
+    numberCommet.innerHTML = `comments (${commentCounter})`;
+  }, 500);
+};
+
+const clearField = (nameInput, commentInput) => {
+  nameInput.value = '';
+  commentInput.value = '';
+};
+
+export {
+  displayDetailCard, displayComments, displayCounterOfComments, clearField,
+};
