+++ src/style.css
@@ -94,7 +94,7 @@ h1 {
 .shows-list {
   padding: 2rem;
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(20rem, 1fr));
+  grid-template-columns: repeat(auto-fit, minmax(25rem, 1fr));
   grid-gap: 4rem;
 }
